chore: デフォルトDB以外を選択できるようにFirebaseの初期化処理を変更した
